+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The Slim script is replaced by a custom livewire solution. This resulted in less
 
 -   Added: `Field::requiredFallbackLocale()` as a alias of `Field::rules('requireFallbackLocale)`.
 -   Added: callback as fourth optional argument for the Select::sync() method that is called after syncing values.
+-   Added: `TableColumnLink::target()` to add a target attribute to the link. Defaults to `_self`.
 -   Removed `Thinktomorrow\Chief\Forms\Fields\Media\MediaType` as a legacy reference to predefined file keys.
 -   Removed `Thinktomorrow\Chief\Forms\Fields\Media\FileDTO` since now `Thinktomorrow\AssetLibrary\Asset` is passed to file field instead.
 -   Removed `Thinktomorrow\Chief\Managers\Assistants\FileUploadAssistant`, `Thinktomorrow\Chief\Managers\Assistants\SlimImageUploadAssistant` and `Thinktomorrow\Chief\Managers\Assistants\RedactorFileUploadAssistant`. These are no longer used and replaced by the Livewire FileComponent upload logic. The latter one was no longer actively used.